Re-chunk consistenly on appends.
When appending, rewrite the last chunk as part of the chunking operation so
that last chunks terminated before a fingerprint will correctly continue into
the new data.

Fix problems related to broken invariants that show up as a result.

Also remove the "seeding" concept of the chunker, it turns out it was broken
(didn't establish the same new state) and isn't necessary.