Fix persisted state of the "origin" tree.
We were reusing the target context for the origin - this was a bad idea
because the NodeContext is supposed to track the commit and change that the
tree was constructed from and we were relying on this when persisting the info
on the origin.  Use a fresh context for the origin tree and update the context
in the replayJournalUntil*() methods.

This change also includes various small fixes to fuse_test and some of the
logging in cache.