Add "traverse" command.
Add the "traverse" comamnd and RPC.  "Traverse" does traversal of the entire
commit/node tree for a branch, causing all objects in the branch to be pulled
to the local store if they are not present and also verified.

This currently only does a full traversal.  There's an "algorithm" indicator
in the API to allow us to do delta traversals.