Implement bidirectional RPC connections.
Add support for bidirectional RPC connections.  WHen a client invokes the
"peerConnected" method, the connection converts to a bidirection channel
allowing the server to make requests back to the client over the same
connection.

This is not yet supported in the peer network code: this change only adds it
to mawfs.rpc in preparation for the latter work.