Don't orphan or release nodes with external references.
This fixes the case where we automatically commit while a file is being
written to.  In that case, during post-commit garbage collection, the file
node was GC'ed and the file was orphaned, causing all subsequent writes to be
ignored.

This change requires distinguishing between a release of a deleted node and
release during GC.