Add better branch divergence check
A branch can subsume another branch even if they both have journals and the
journals are based on different commits, as long as a) one of those commits
is derived from the other and b) the journal based on the older commit is
subsumed by the next commit on the path to the newer commit.

This change fully implements this check, so from now on we should only
conclude that the branches have diverged (and that we have to merge) if the
commit history diverges or the journals diverge (either from one another or
if the journal based off the older commit diverges from the subsequent commit).