Fix corrupt reads of large files (and "seek")
The fuse API wasn't doing "seek" on read, resulting in sporadic failures
during reads of large files (not to mention the general inability to do seek()
on any reads).

This change also removes the merge code from the push test, which is no longer
needed (this should have gone into the last change).